Enable execution of application functions from command line
The basic functions of the JS16_Bootloader application could only be run from the GUI.
A console mode execution is often desired to automate tasks, for example in test systems.

Where required, GUI-related code blocks are wrapped inside an if block that evaluates to true when running in GUI mode.
This enabled sharing of related functions between both GUI and command line modes.