Don't assume Windows build when $(OS) is not set
commit add56e0fbb47 ("V4.12.1.270 - Updated to using MinGW-64 for all
windows buld and libraries") set the OS variable to "Windows_NT" when it
was unset. However, the OS variable doesn't exist at all on Linux and so
is always unset there. Obviously, we don't want to assume a Windows
build on a Linux system, so remove the new logic to revert back to the
old behavior.

Fixes: add56e0fbb47 ("V4.12.1.270 - Updated to using MinGW-64 for all windows buld and libraries")